Refactor UI and Storage for improved OOP structure
